backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==9.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.8.1.78
mtcnn==0.1.1
keras-facenet==0.3.2
pydantic==2.5.0
pillow==10.1.0
cryptography==41.0.7